Add input fields for modifying matrix elements
Currently does not do anything, in the future this will propagate the
changes to projection and potentially give the user the option to
instead propagate to subscreen (rarer usecase). Fixed TabletDisplay
displaying the incorrect position after resizing but not moving the
projection. Check HTML5 validation for input fields in BoundForm, and
allow invalid inputs to remain there but don't propagated them.