Update Huion 420 working area
Huion published an incorrect working area as well as published a doubled
LPI. It is not clear if this was a mistake or not, however I suspect
they were just trying to make the number higher than Wacom's very
reasonable 2540 LPI (100 L/mm). From my calculations as well as hawku's
it appears the 420 has a working area of 4" x 2.34" at 2000 LPI. Also,
hard code the 1080p play area.